+++ cmd/emp/dev.go
@@ -1,7 +1,5 @@
-// +build !release
-
 package main
 
 const (
-	Version = "0.9.2"
+	Version = "0.10.0"
 )