+++ empire.go
@@ -1,7 +1,6 @@
 package empire // import "github.com/remind101/empire"
 
 import (
-	"fmt"
 	"io"
 	"io/ioutil"
 
@@ -562,14 +561,9 @@ func PullAndExtract(c *dockerutil.Client) ProcfileExtractor {
 	)
 
 	return ProcfileExtractor(func(ctx context.Context, img image.Image, w io.Writer) (procfile.Procfile, error) {
-		repo := img.Repository
-		if img.Registry != "" {
-			repo = fmt.Sprintf("%s/%s", img.Registry, img.Repository)
-		}
-
 		if err := c.PullImage(ctx, docker.PullImageOptions{
 			Registry:      img.Registry,
-			Repository:    repo,
+			Repository:    img.Repository,
 			Tag:           img.Tag,
 			OutputStream:  w,
 			RawJSONStream: true,